+++ src/sdk.test.ts
@@ -2,6 +2,7 @@ import { SorobanDomainsSDK } from "./sdk.ts";
 import { describe, test } from "@std/testing/bdd";
 import { assertEquals } from "@std/assert";
 import { decodeHex } from "@std/encoding";
+import { type Record, RecordType } from "./types.ts";
 
 describe("Basic logic", (): void => {
   test("The `hash` function", () => {
@@ -23,15 +24,27 @@ describe("Basic logic", (): void => {
   });
 
   test("It should validate domains correctly", (): void => {
-    assertEquals(true, SorobanDomainsSDK.isValidDomain('stellar.xlm'));
-    assertEquals(true, SorobanDomainsSDK.isValidDomain('dev.stellar.xlm'));
-    assertEquals(false, SorobanDomainsSDK.isValidDomain('another.dev.stellar.xlm'));
-    assertEquals(false, SorobanDomainsSDK.isValidDomain('stellar'));
-    assertEquals(false, SorobanDomainsSDK.isValidDomain('stellar..xlm'));
-    assertEquals(false, SorobanDomainsSDK.isValidDomain(' stellar.xlm'));
-    assertEquals(false, SorobanDomainsSDK.isValidDomain('qwertyuiopasdfghjklzxcvbnm.xlm'));
-    assertEquals(false, SorobanDomainsSDK.isValidDomain('hello-world.xlm'));
-    assertEquals(false, SorobanDomainsSDK.isValidDomain('steLLar.xlm'));
-    assertEquals(false, SorobanDomainsSDK.isValidDomain('ste11ar.xlm'));
+    assertEquals(true, SorobanDomainsSDK.isValidDomain("stellar.xlm"));
+    assertEquals(true, SorobanDomainsSDK.isValidDomain("dev.stellar.xlm"));
+    assertEquals(false, SorobanDomainsSDK.isValidDomain("another.dev.stellar.xlm"));
+    assertEquals(false, SorobanDomainsSDK.isValidDomain("stellar"));
+    assertEquals(false, SorobanDomainsSDK.isValidDomain("stellar..xlm"));
+    assertEquals(false, SorobanDomainsSDK.isValidDomain(" stellar.xlm"));
+    assertEquals(false, SorobanDomainsSDK.isValidDomain("qwertyuiopasdfghjklzxcvbnm.xlm"));
+    assertEquals(false, SorobanDomainsSDK.isValidDomain("hello-world.xlm"));
+    assertEquals(false, SorobanDomainsSDK.isValidDomain("steLLar.xlm"));
+    assertEquals(false, SorobanDomainsSDK.isValidDomain("ste11ar.xlm"));
+  });
+});
+
+describe("Connection with public registry", () => {
+  test("Search an existing domain", async () => {
+    const now: number = performance.now();
+    const sdk: SorobanDomainsSDK = new SorobanDomainsSDK({ rpcUrl: "https://mainnet.sorobanrpc.com" });
+    const domain: Record = await sdk.searchDomain({ domain: "tomer" });
+    assertEquals(domain.type, RecordType.Domain);
+    const then: number = performance.now();
+
+    console.log(`${(then - now).toFixed(2)}ms`);
   });
 });