+++ src/sdk.ts
@@ -1,9 +1,13 @@
 import {
   type Domain,
   type DomainStorageValue,
+  KeyValueDbContract,
   type Record,
   RecordKey,
   RecordType,
+  RegistryContract,
+  ReverseRegistrarContract,
+  SIMULATION_ACCOUNT,
   type SorobanDomainsSDKParams,
   type SubDomain,
 } from "./types.ts";
@@ -11,11 +15,47 @@ import { crypto } from "@std/crypto";
 import { encodeHex } from "@std/encoding";
 import { concat } from "@std/bytes";
 import { Domain404Error, DomainData404Error, DomainDataUnsupportedValueType, ReverseDomain404Error } from "./errors.ts";
-import type { Account, Contract, rpc, Transaction, xdr } from "@stellar/stellar-sdk";
+import {
+  Account,
+  Contract,
+  nativeToScVal,
+  Networks,
+  rpc,
+  scValToNative,
+  type Transaction,
+  TransactionBuilder,
+  xdr,
+} from "@stellar/stellar-sdk";
 import { decodeHex } from "@std/encoding/hex";
 
 export class SorobanDomainsSDK {
-  constructor(public global: SorobanDomainsSDKParams) {}
+  readonly #rpcUrl?: string;
+  readonly #simulationAccount: string;
+  readonly #globalFee: string;
+  readonly #registerContract: string;
+  readonly #keyValueDbContract: string;
+  readonly #reverseRegistrarContract: string;
+  readonly #network: Networks;
+  readonly #defaultTimeout: number;
+
+  get server(): rpc.Server {
+    if (!this.#rpcUrl) {
+      throw new Error("This method requires that you define an `rpcUrl` value.");
+    }
+
+    return new rpc.Server(this.#rpcUrl);
+  }
+
+  constructor(params: SorobanDomainsSDKParams) {
+    this.#rpcUrl = params.rpcUrl;
+    this.#simulationAccount = params.simulationAccount || SIMULATION_ACCOUNT;
+    this.#globalFee = params.defaultFee || "100";
+    this.#registerContract = params.registryContractId || RegistryContract.v0;
+    this.#keyValueDbContract = params.keyValuesDatabaseContractId || KeyValueDbContract.v0;
+    this.#reverseRegistrarContract = params.reverseRegistrarContractId || ReverseRegistrarContract.v0;
+    this.#network = params.network || Networks.PUBLIC;
+    this.#defaultTimeout = params.defaultTimeout || 0;
+  }
 
   static hash(text: string | Uint8Array): Uint8Array {
     return new Uint8Array(
@@ -51,9 +91,9 @@ export class SorobanDomainsSDK {
    * @param domain {String} - The domain to validate, for example: stellar.xlm
    */
   static isValidDomain(domain: string): boolean {
-    const domainRegex: RegExp = new RegExp('^[a-z]+(\\.[a-z]+)*\\.[a-z]{2,}$');
+    const domainRegex: RegExp = new RegExp("^[a-z]+(\\.[a-z]+)*\\.[a-z]{2,}$");
     if (!domainRegex.test(domain)) return false;
-    const parts: string[] = domain.split('.');
+    const parts: string[] = domain.split(".");
     if (parts.length > 3) return false;
     return parts.every((part: string): boolean => part.length <= 15);
   }
@@ -63,40 +103,34 @@ export class SorobanDomainsSDK {
    * This function doesn't validate the domain you're providing is a valid one.
    */
   async searchDomain(params: { domain: string; subDomain?: string }): Promise<Record> {
-    if (!this.global.vaultsContractId) {
-      throw new Error(`Vault's contract id was not provided`);
-    }
-
     const domainNode: string = SorobanDomainsSDK.parseDomain({
       domain: params.domain.toLocaleLowerCase(),
       subDomain: params.subDomain?.toLocaleLowerCase(),
     });
 
-    const contract: Contract = new this.global.stellarSDK.Contract(this.global.vaultsContractId);
-    const nodeBytes: xdr.ScVal = this.global.stellarSDK.xdr.ScVal.scvBytes(decodeHex(domainNode));
+    const contract: Contract = new Contract(this.#registerContract);
+    const nodeBytes: xdr.ScVal = xdr.ScVal.scvBytes(decodeHex(domainNode));
 
-    const record_key: xdr.ScVal = this.global.stellarSDK.xdr.ScVal.scvVec([
-      this.global.stellarSDK.xdr.ScVal.scvSymbol(params.subDomain ? RecordKey.SubRecord : RecordKey.Record),
+    const record_key: xdr.ScVal = xdr.ScVal.scvVec([
+      xdr.ScVal.scvSymbol(params.subDomain ? RecordKey.SubRecord : RecordKey.Record),
       nodeBytes,
     ]);
 
-    const account: Account = await this.global.rpc.getAccount(this.global.simulationAccount);
-
-    const transaction: Transaction = new this.global.stellarSDK.TransactionBuilder(account, {
-      networkPassphrase: this.global.network,
-      fee: this.global.defaultFee,
+    const transaction: Transaction = new TransactionBuilder(new Account(this.#simulationAccount, "0"), {
+      networkPassphrase: this.#network,
+      fee: this.#globalFee,
     })
-      .setTimeout(this.global.defaultTimeout || 0)
+      .setTimeout(this.#defaultTimeout)
       .addOperation(contract.call("record", record_key))
       .build();
 
-    const sim: rpc.Api.SimulateTransactionResponse = await this.global.rpc.simulateTransaction(transaction);
+    const sim: rpc.Api.SimulateTransactionResponse = await this.server.simulateTransaction(transaction);
 
-    if (this.global.stellarSDK.rpc.Api.isSimulationError(sim)) {
+    if (rpc.Api.isSimulationError(sim)) {
       throw new Error(sim.error);
     }
 
-    const result = this.global.stellarSDK.scValToNative(sim.result!.retval);
+    const result = scValToNative(sim.result!.retval);
 
     if (!result) {
       throw new Domain404Error();
@@ -128,31 +162,25 @@ export class SorobanDomainsSDK {
   }
 
   async getDomainData(params: { node: string; key: string }): Promise<DomainStorageValue> {
-    if (!this.global.valuesDatabaseContractId) {
-      throw new Error(`KeyValue Database contract id was not provided`);
-    }
+    const contract: Contract = new Contract(this.#keyValueDbContract);
+    const nodeBytes: xdr.ScVal = xdr.ScVal.scvBytes(decodeHex(params.node));
+    const keySymbol: xdr.ScVal = xdr.ScVal.scvSymbol(params.key);
 
-    const contract: Contract = new this.global.stellarSDK.Contract(this.global.valuesDatabaseContractId);
-    const nodeBytes: xdr.ScVal = this.global.stellarSDK.xdr.ScVal.scvBytes(decodeHex(params.node));
-    const keySymbol: xdr.ScVal = this.global.stellarSDK.xdr.ScVal.scvSymbol(params.key);
-
-    const account: Account = await this.global.rpc.getAccount(this.global.simulationAccount);
-
-    const transaction: Transaction = new this.global.stellarSDK.TransactionBuilder(account, {
-      networkPassphrase: this.global.network,
-      fee: this.global.defaultFee,
+    const transaction: Transaction = new TransactionBuilder(new Account(this.#simulationAccount, "0"), {
+      networkPassphrase: this.#network,
+      fee: this.#globalFee,
     })
-      .setTimeout(this.global.defaultTimeout || 0)
+      .setTimeout(this.#defaultTimeout)
       .addOperation(contract.call("get", nodeBytes, keySymbol))
       .build();
 
-    const sim: rpc.Api.SimulateTransactionResponse = await this.global.rpc.simulateTransaction(transaction);
+    const sim: rpc.Api.SimulateTransactionResponse = await this.server.simulateTransaction(transaction);
 
-    if (this.global.stellarSDK.rpc.Api.isSimulationError(sim)) {
+    if (rpc.Api.isSimulationError(sim)) {
       throw new Error(sim.error);
     }
 
-    const result = this.global.stellarSDK.scValToNative(sim.result!.retval);
+    const result = scValToNative(sim.result!.retval);
 
     if (!result) {
       throw new DomainData404Error();
@@ -165,34 +193,30 @@ export class SorobanDomainsSDK {
     tx: Transaction;
     sim: rpc.Api.SimulateTransactionRestoreResponse | rpc.Api.SimulateTransactionSuccessResponse;
   }> {
-    if (!this.global.valuesDatabaseContractId) {
-      throw new Error(`KeyValue Database contract id was not provided`);
-    }
-
-    const contract: Contract = new this.global.stellarSDK.Contract(this.global.valuesDatabaseContractId);
-    const nodeBytes: xdr.ScVal = this.global.stellarSDK.xdr.ScVal.scvBytes(decodeHex(params.node));
-    const keySymbol: xdr.ScVal = this.global.stellarSDK.xdr.ScVal.scvSymbol(params.key);
+    const contract: Contract = new Contract(this.#keyValueDbContract);
+    const nodeBytes: xdr.ScVal = xdr.ScVal.scvBytes(decodeHex(params.node));
+    const keySymbol: xdr.ScVal = xdr.ScVal.scvSymbol(params.key);
     let value: xdr.ScVal;
 
     switch (params.value[0]) {
       case "Bytes":
-        value = this.global.stellarSDK.xdr.ScVal.scvVec([
-          this.global.stellarSDK.xdr.ScVal.scvSymbol("Bytes"),
-          this.global.stellarSDK.xdr.ScVal.scvBytes(params.value[1]),
+        value = xdr.ScVal.scvVec([
+          xdr.ScVal.scvSymbol("Bytes"),
+          xdr.ScVal.scvBytes(params.value[1]),
         ]);
         break;
 
       case "Number":
-        value = this.global.stellarSDK.xdr.ScVal.scvVec([
-          this.global.stellarSDK.xdr.ScVal.scvSymbol("Number"),
-          this.global.stellarSDK.nativeToScVal(params.value[1], { type: "i128" }),
+        value = xdr.ScVal.scvVec([
+          xdr.ScVal.scvSymbol("Number"),
+          nativeToScVal(params.value[1], { type: "i128" }),
         ]);
         break;
 
       case "String":
-        value = this.global.stellarSDK.xdr.ScVal.scvVec([
-          this.global.stellarSDK.xdr.ScVal.scvSymbol("String"),
-          this.global.stellarSDK.xdr.ScVal.scvString(params.value[1]),
+        value = xdr.ScVal.scvVec([
+          xdr.ScVal.scvSymbol("String"),
+          xdr.ScVal.scvString(params.value[1]),
         ]);
         break;
 
@@ -200,24 +224,24 @@ export class SorobanDomainsSDK {
         throw new DomainDataUnsupportedValueType();
     }
 
-    const account: Account = await this.global.rpc.getAccount(params.source);
+    const account: Account = await this.server.getAccount(params.source);
 
-    const transaction: Transaction = new this.global.stellarSDK.TransactionBuilder(account, {
-      networkPassphrase: this.global.network,
-      fee: this.global.defaultFee,
+    const transaction: Transaction = new TransactionBuilder(account, {
+      networkPassphrase: this.#network,
+      fee: this.#globalFee,
     })
-      .setTimeout(this.global.defaultTimeout || 0)
+      .setTimeout(this.#defaultTimeout)
       .addOperation(contract.call("set", nodeBytes, keySymbol, value))
       .build();
 
-    const sim: rpc.Api.SimulateTransactionResponse = await this.global.rpc.simulateTransaction(transaction);
+    const sim: rpc.Api.SimulateTransactionResponse = await this.server.simulateTransaction(transaction);
 
-    if (this.global.stellarSDK.rpc.Api.isSimulationError(sim)) {
+    if (rpc.Api.isSimulationError(sim)) {
       throw new Error(sim.error);
     }
 
     return {
-      tx: this.global.stellarSDK.rpc.assembleTransaction(transaction, sim).build(),
+      tx: rpc.assembleTransaction(transaction, sim).build(),
       sim,
     };
   }
@@ -226,32 +250,28 @@ export class SorobanDomainsSDK {
     tx: Transaction;
     sim: rpc.Api.SimulateTransactionRestoreResponse | rpc.Api.SimulateTransactionSuccessResponse;
   }> {
-    if (!this.global.valuesDatabaseContractId) {
-      throw new Error(`KeyValue Database contract id was not provided`);
-    }
+    const contract: Contract = new Contract(this.#keyValueDbContract);
+    const nodeBytes: xdr.ScVal = xdr.ScVal.scvBytes(decodeHex(params.node));
+    const keySymbol: xdr.ScVal = xdr.ScVal.scvSymbol(params.key);
 
-    const contract: Contract = new this.global.stellarSDK.Contract(this.global.valuesDatabaseContractId);
-    const nodeBytes: xdr.ScVal = this.global.stellarSDK.xdr.ScVal.scvBytes(decodeHex(params.node));
-    const keySymbol: xdr.ScVal = this.global.stellarSDK.xdr.ScVal.scvSymbol(params.key);
+    const account: Account = await this.server.getAccount(params.source);
 
-    const account: Account = await this.global.rpc.getAccount(params.source);
-
-    const transaction: Transaction = new this.global.stellarSDK.TransactionBuilder(account, {
-      networkPassphrase: this.global.network,
-      fee: this.global.defaultFee,
+    const transaction: Transaction = new TransactionBuilder(account, {
+      networkPassphrase: this.#network,
+      fee: this.#globalFee,
     })
-      .setTimeout(this.global.defaultTimeout || 0)
+      .setTimeout(this.#defaultTimeout)
       .addOperation(contract.call("remove", nodeBytes, keySymbol))
       .build();
 
-    const sim: rpc.Api.SimulateTransactionResponse = await this.global.rpc.simulateTransaction(transaction);
+    const sim: rpc.Api.SimulateTransactionResponse = await this.server.simulateTransaction(transaction);
 
-    if (this.global.stellarSDK.rpc.Api.isSimulationError(sim)) {
+    if (rpc.Api.isSimulationError(sim)) {
       throw new Error(sim.error);
     }
 
     return {
-      tx: this.global.stellarSDK.rpc.assembleTransaction(transaction, sim).build(),
+      tx: rpc.assembleTransaction(transaction, sim).build(),
       sim,
     };
   }
@@ -277,13 +297,9 @@ export class SorobanDomainsSDK {
     tx: Transaction;
     sim: rpc.Api.SimulateTransactionRestoreResponse | rpc.Api.SimulateTransactionSuccessResponse;
   }> {
-    if (!this.global.reverseRegistrarContractId) {
-      throw new Error(`Reverse Registrar contract id was not provided`);
-    }
+    const addressScval = nativeToScVal(params.address, { type: "address" });
 
-    const addressScval = this.global.stellarSDK.nativeToScVal(params.address, { type: "address" });
-
-    let domainScval = this.global.stellarSDK.xdr.ScVal.scvVoid();
+    let domainScval = xdr.ScVal.scvVoid();
     if (params.domain !== null) {
       const parts = params.domain.toLocaleLowerCase().split(".");
       if (parts.length < 2) {
@@ -294,7 +310,7 @@ export class SorobanDomainsSDK {
         sld: new TextEncoder().encode(parts[parts.length - 2]),
         subs: parts.slice(0, parts.length - 2).map((part) => new TextEncoder().encode(part)),
       };
-      domainScval = this.global.stellarSDK.nativeToScVal(domainParts, {
+      domainScval = nativeToScVal(domainParts, {
         type: {
           tld: ["symbol"],
           sld: ["symbol"],
@@ -303,24 +319,24 @@ export class SorobanDomainsSDK {
       });
     }
 
-    const contract: Contract = new this.global.stellarSDK.Contract(this.global.reverseRegistrarContractId);
-    const account: Account = await this.global.rpc.getAccount(params.source);
-    const transaction: Transaction = new this.global.stellarSDK.TransactionBuilder(account, {
-      networkPassphrase: this.global.network,
-      fee: this.global.defaultFee,
+    const contract: Contract = new Contract(this.#reverseRegistrarContract);
+    const account: Account = await this.server.getAccount(params.source);
+    const transaction: Transaction = new TransactionBuilder(account, {
+      networkPassphrase: this.#network,
+      fee: this.#globalFee,
     })
-      .setTimeout(this.global.defaultTimeout || 0)
+      .setTimeout(this.#defaultTimeout)
       .addOperation(contract.call("set", addressScval, domainScval))
       .build();
 
-    const sim: rpc.Api.SimulateTransactionResponse = await this.global.rpc.simulateTransaction(transaction);
+    const sim: rpc.Api.SimulateTransactionResponse = await this.server.simulateTransaction(transaction);
 
-    if (this.global.stellarSDK.rpc.Api.isSimulationError(sim)) {
+    if (rpc.Api.isSimulationError(sim)) {
       throw new Error(sim.error);
     }
 
     return {
-      tx: this.global.stellarSDK.rpc.assembleTransaction(transaction, sim).build(),
+      tx: rpc.assembleTransaction(transaction, sim).build(),
       sim,
     };
   }
@@ -338,29 +354,24 @@ export class SorobanDomainsSDK {
    * @throws ReverseDomain404Error if no reverse domain is set for the address
    */
   async getReverseDomain(address: string): Promise<string> {
-    if (!this.global.reverseRegistrarContractId) {
-      throw new Error(`Reverse Registrar contract id was not provided`);
-    }
+    const contract: Contract = new Contract(this.#reverseRegistrarContract);
+    const addressScval: xdr.ScVal = nativeToScVal(address, { type: "address" });
 
-    const contract: Contract = new this.global.stellarSDK.Contract(this.global.reverseRegistrarContractId);
-    const addressScval: xdr.ScVal = this.global.stellarSDK.nativeToScVal(address, { type: "address" });
-
-    const account: Account = await this.global.rpc.getAccount(this.global.simulationAccount);
-    const transaction: Transaction = new this.global.stellarSDK.TransactionBuilder(account, {
-      networkPassphrase: this.global.network,
-      fee: this.global.defaultFee,
+    const transaction: Transaction = new TransactionBuilder(new Account(this.#simulationAccount, "0"), {
+      networkPassphrase: this.#network,
+      fee: this.#globalFee,
     })
-      .setTimeout(this.global.defaultTimeout || 0)
+      .setTimeout(this.#defaultTimeout)
       .addOperation(contract.call("get", addressScval))
       .build();
 
-    const sim: rpc.Api.SimulateTransactionResponse = await this.global.rpc.simulateTransaction(transaction);
+    const sim: rpc.Api.SimulateTransactionResponse = await this.server.simulateTransaction(transaction);
 
-    if (this.global.stellarSDK.rpc.Api.isSimulationError(sim)) {
+    if (rpc.Api.isSimulationError(sim)) {
       throw new Error(sim.error);
     }
 
-    const result = this.global.stellarSDK.scValToNative(sim.result!.retval);
+    const result = scValToNative(sim.result!.retval);
 
     if (!result) {
       throw new ReverseDomain404Error();